+++ components/org.wso2.carbon.identity.workflow.engine/src/main/java/org/wso2/carbon/identity/workflow/engine/exception/WorkflowEngineRuntimeException.java
@@ -1,7 +1,6 @@
 package org.wso2.carbon.identity.workflow.engine.exception;
 
 import org.wso2.carbon.database.utils.jdbc.exceptions.DataAccessException;
-import org.wso2.carbon.identity.workflow.mgt.exception.InternalWorkflowException;
 
 /**
  * Exception class to handle runtime exceptions.
